# Conflicts:
#	src/main/java/com/nhsd/a2si/capacityinformation/domain/CapacityInformation.java

+++ src/test/java/com/nhsd/a2si/capacityinformation/domain/CapacityInformationTest.java
@@ -96,7 +96,7 @@ public class CapacityInformationTest {
 
         capacityInformation.setNumberOfPeopleWaiting(peopleWaiting);
 
-        assertEquals(26, capacityInformation.getNumberOfPeopleWaiting());
+        assertEquals(26, capacityInformation.getNumberOfPeopleWaiting().intValue());
     }
